Current State

+++ fileforge-std/src/magic/mod.rs
@@ -1,8 +1,10 @@
-use error::{error::MagicError, invalid::MagicInvalid};
+use error::invalid::MagicInvalid;
 use fileforge::{
   binary_reader::{readable::Readable, BinaryReader, PrimitiveReader},
   stream::ReadableStream,
 };
+
+use crate::magic::error::MagicError;
 
 pub mod error;
 pub mod renderable;
@@ -28,7 +30,7 @@ impl<'pool, const SIZE: usize, S: ReadableStream<Type = u8>> Readable<'pool, S> 
   async fn read(reader: &mut BinaryReader<'pool, S>, argument: Self::Argument) -> Result<Self, Self::Error> {
     let content = Self::from_bytes(reader.get::<[u8; SIZE]>().await?);
 
-    MagicInvalid::assert(content, argument, || reader.create_physical_diagnostic(-(SIZE as i64), Some(SIZE as u64), "Magic"))?;
+    MagicInvalid::assert(content, argument, || reader.create_physical_diagnostic(-(SIZE as i128), Some(SIZE as u64), "Magic"))?;
 
     Ok(content)
   }
